old-tests/submission/Opera/script_scheduling/091.html

<!DOCTYPE html>
<html><head>
	<title> scheduler: force-async off on non-parser-inserted script</title>
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="testlib/testlib.js"></script>
</head>
<body>

	<div id="log">FAILED (This TC requires JavaScript enabled)</div>

	<script>
          var t = async_test(undefined, {timeout:4000});
          
          sources = ["scripts/delay.php?time=2&url=include-1.js", "scripts/delay.php?time=1&url=include-2.js"];
          sources.forEach(function(x) {
            var script = document.createElement("script");
            script.src = x;
            t.step(function() {assert_equals(script.async, true, "async IDL attribute on script creation")});
            script.async = false;
            t.step(function() {assert_equals(script.async, false, "async IDL attribute after setting")});
            t.step(function() {assert_equals(script.getAttribute("async"), null, "async content attribute after setting")});
            document.head.appendChild(script);
          });

	  onload = t.step_func(function() {
                                 assert_array_equals(eventOrder, ['external script #1', 'external script #2']);
                                 t.done();
                             });
        </script>
</body></html>
